spike: cr feedback + amending sentry init packaging for webpack issue
