+++ packages/plugin-core/src/utils/analytics.ts
@@ -1,6 +1,7 @@
 import { getOS, SegmentClient } from "@dendronhq/common-server";
-import { DendronExtension } from "../workspace";
+import * as Sentry from "@sentry/node";
 import * as vscode from "vscode";
+import { DendronExtension } from "../workspace";
 
 export type SegmentContext = Partial<{
   app: Partial<{ name: string; version: string; build: string }>;
@@ -52,4 +53,24 @@ export class AnalyticsUtils {
       userAgent: vscode.env.appName,
     };
   }
+}
+
+/**
+ * Wraps a callback function with a try/catch block.  In the catch, any
+ * exceptions that were encountered will be uploaded to Sentry and then
+ * rethrown.
+ * @param callback the function to wrap
+ * @returns the wrapped callback function
+ */
+export function sentryReportingCallback(
+  callback: (...args: any[]) => any
+): (...args: any[]) => any {
+  return (args) => {
+    try {
+      return callback(args);
+    } catch (error) {
+      Sentry.captureException(error);
+      throw error;
+    }
+  };
 }
